♻️ Split analyzer into block/graph/workflow modules
Move analyzer constants into types package. Update expr types,
bootstrap definitions, LangGraph codegen, and tests.

Made-with: Cursor
