+++ compiler/codegen/codegen.go
@@ -6,6 +6,7 @@ import (
 	"github.com/thakee/orca/compiler/analyzer"
 	"github.com/thakee/orca/compiler/ast"
 	"github.com/thakee/orca/compiler/diagnostic"
+	"github.com/thakee/orca/compiler/types"
 )
 
 // CodegenOutput holds the complete output from a code generation backend.
@@ -74,7 +75,7 @@ func (b *BaseBackend) CollectBlocksByKind(kind string) []*ast.BlockStatement {
 
 // CollectLets returns all let block statements.
 func (b *BaseBackend) CollectLets() []*ast.BlockStatement {
-	return b.CollectBlocksByKind(analyzer.BlockKindLet)
+	return b.CollectBlocksByKind(types.BlockKindLet)
 }
 
 // BlockByName returns the block statement with the given name, or nil if not found.
