+++ compiler/diagnostic/diagnostic.go
@@ -40,10 +40,7 @@ const (
 	CodeUnexpectedExpr   = "unexpected-expr"       // expression not allowed in this context
 	CodeInvalidWorkNode  = "invalid-workflow-node" // block kind not allowed as workflow node
 	CodeTriggerAsTarget  = "trigger-as-target"     // trigger block used as edge target instead of source
-	CodeAmbiguousStart   = "ambiguous-start"       // multiple entry nodes without triggers to disambiguate
-	CodeDanglingEntry    = "dangling-entry"        // entry node unreachable because no trigger points to it
 	CodeCyclicDependency = "cyclic-dependency"     // blocks form a dependency cycle
-	CodeBranchOutgoing = "branch-outgoing" // branch block used on left side of arrow
 )
 
 // Diagnostic represents a single compiler message (error, warning, etc.)
